Split build_pile.py into pilebuild/, and test the vg_scale passes
build_pile.py was 1885 lines interleaving five dataset loaders, provenance
and fingerprinting, and six read-only audit modes, with a 304-line
_load_vg_scale in the middle of it.

The CLI and the per-cell build loop stay in build_pile.py; everything a
build does inside itself moves to pilebuild/, one module per question. The
names sibling scripts and tests import off build_pile (_vg_image_paths,
_band_categories, build_cell, rebuildable, ...) are re-exported, so the six
make_*.py sheet builders and both build_arm.py callers are untouched.

Two changes are more than a move:

One module per dataset kind, owning both halves. `load()` (how the cell is
built) and `check()` (what a rebuild of it reads) now live together and are
dispatched through one registry keyed by DATASETS[ds]["kind"], replacing two
kind switches that sat a thousand lines apart. That drift is #3299: the
rebuild canary checked COCO_IMAGES while the builder opened val2017.zip
inline, and reported coco_val REBUILD-BROKEN against a staging area that was
entirely intact. An unregistered kind now fails at dispatch instead of
falling through to the demo loader, which had made a typo in DATASETS build
a plausible-looking cell out of the wrong source.

_load_vg_scale becomes six named passes. read_vg_labels, anchor_to_coco,
apply_corrections, band_candidates, designate_cells and draw_negatives each
take what they read and return what they produce, so they run without the VG
source. That is what the split is for: apply_corrections is the single point
at which a box crosses from normalised into pixel space (#3281, invisible
downstream because the band is derived from the same corrupted box), and
designate_cells decides whether a rebuild keeps the images a human already
reviewed (two regressions there orphaned 49 of 360 negatives and 99 of 360
boxed positives, both behind cells that looked perfect). Neither had a test,
because neither was callable.

tests_lib/core/test_pile_vg_scale.py covers them: the normalised-to-pixel
round trip is exact, anchoring replaces rather than merges and skips
re-framed copies, scattered instances leave every band, selection moves at
most one seat when the candidate pool changes, reviewed images outrank
unreviewed ones, and roster entries that stop being eligible drop out and
backfill by rank. test_pile_loaders.py pins the registry: a kind in
DATASETS with no module, or a module missing either half, fails here rather
than on the GRID.

pyrightconfig.json gains the pile directory on extraPaths so those tests
type-check against the real modules; scripts/ stays excluded from checking.

Closes #3410
