+++ scripts/experiments/pile/build_pile.py
@@ -21,6 +21,13 @@ Usage::
     python build_pile.py --provenance                # which device built each cell
     python build_pile.py --backfill-provenance       # fingerprint the pre-#3160 cells
 
+This module is the CLI and the per-cell build loop; everything it does inside a
+build lives in :mod:`pilebuild`, split by the question each part answers -- one
+loader module per ``DATASETS[ds]["kind"]``, and separate modules for provenance,
+geometry, the manifest and the read-only audit modes. The names below are
+re-exported so that ``import build_pile`` stays the entry point every sibling
+script and test already uses.
+
 ``--rebuildable`` is the complement to ``--verify``, and the two answer
 different questions. ``--verify`` asks whether the cells on disk are usable;
 ``--rebuildable`` asks whether they could be produced again. A cell that loads
@@ -40,935 +47,52 @@ with it (#3281).
 from __future__ import annotations
 
 import argparse
-import gzip
-import hashlib
-import json
 import os
-import sys
 import time
-import zipfile
-from collections import defaultdict
 from contextlib import contextmanager
-from pathlib import Path
 
 import pile_config as pc
 
 pc.setup_env()
 
-
-def log(msg: str) -> None:
-    print(f"[pile] {msg}", flush=True)
-
-
-def assert_vtscore_is_this_checkout() -> None:
-    """Refuse to run against a different checkout's ``vtscore``.
-
-    The venv's editable install points at the main checkout. If anything
-    resolves ``vtscore`` there instead of here, cells get embedded by whatever
-    code that tree happens to be on — silently, and possibly by a different
-    embedder implementation. Cheap to assert, expensive to discover later.
-    """
-    import vtscore  # noqa: PLC0415
-
-    want = Path(__file__).resolve().parents[3]
-    got = Path(vtscore.__file__).resolve().parent.parent
-    if got != want:
-        raise SystemExit(
-            f"vtscore resolved to {got}, not this checkout ({want}).\n"
-            f"  Something put another checkout ahead on sys.path — usually the venv's\n"
-            f"  editable install. Re-run with VTS_REPO={want} set for THIS command\n"
-            f"  (note `VAR=x cmd1 && cmd2` sets VAR for cmd1 only)."
-        )
-
-
-def _calibration_path() -> None:
-    calib = Path(__file__).resolve().parent.parent / "calibration"
-    if str(calib) not in sys.path:
-        sys.path.insert(0, str(calib))
-
-
-def _cells_io():
-    """Import the calibration harness's pickle IO (drops bytes, keeps patch_grid)."""
-    _calibration_path()
-    import _cells_io  # noqa: PLC0415
-
-    return _cells_io
-
-
-def _experiment_config():
-    """Import the calibration harness's category-selection config."""
-    _calibration_path()
-    import experiment_config  # noqa: PLC0415
-
-    return experiment_config
-
-
-# --------------------------------------------------------------------------
-# COCO: not a demo dataset, so assemble medias from the staged zip + annotations
-# --------------------------------------------------------------------------
-
-
-def _coco_annotations() -> tuple[dict[int, list[dict]], dict[int, str]]:
-    """``({image_id: [{box, label}, ...]}, {image_id: file_name})``.
-
-    Mirrors ``calibration/build_coco_pickle.py``: boxes already normalised to
-    [0, 1], ``iscrowd`` regions kept (they are still true instances of the
-    category, and positives are defined by category presence).
-    """
-    regions: dict[int, list[dict]] = defaultdict(list)
-    filenames: dict[int, str] = {}
-    with gzip.open(pc.COCO_ANNOTATIONS, "rt") as fh:
-        for line in fh:
-            row = json.loads(line)
-            image_id = int(row["image_id"])
-            filenames[image_id] = row["file_name"]
-            regions[image_id].append(
-                {
-                    "box": [float(row["x0"]), float(row["y0"]), float(row["x1"]), float(row["y1"])],
-                    "label": row["name"],
-                }
-            )
-    return dict(regions), filenames
-
-
-def _load_coco(medias: dict[int, dict], embedder_name: str) -> None:
-    """Populate *medias* with COCO-val images (bytes read straight from the zip).
-
-    Only images the annotation file covers are kept: an image with no category
-    can be neither a positive nor a meaningful negative.
-    """
-    if not pc.COCO_ANNOTATIONS.exists():
-        raise SystemExit(f"missing COCO annotations: {pc.COCO_ANNOTATIONS}")
-    zip_path = pc.COCO_VAL_ZIP
-    if not zip_path.exists():
-        raise SystemExit(f"missing COCO images zip: {zip_path}")
-
-    regions_by_image, filenames = _coco_annotations()
-    log(f"  coco: {len(regions_by_image)} annotated images")
-
-    with zipfile.ZipFile(zip_path) as zf:
-        members = {Path(n).name: n for n in zf.namelist() if n.endswith(".jpg")}
-        missing = 0
-        for image_id in sorted(regions_by_image):
-            regions = regions_by_image[image_id]
-            fname = Path(filenames[image_id]).name
-            member = members.get(fname)
-            if member is None:
-                missing += 1
-                continue
-            counts: dict[str, int] = defaultdict(int)
-            for r in regions:
-                counts[r["label"]] += 1
-            # Most-annotated first, so ``category`` is the dominant object.
-            ordered = [c for c, _ in sorted(counts.items(), key=lambda kv: (-kv[1], kv[0]))]
-            medias[image_id] = {
-                "id": image_id,
-                "media_type": "image",
-                "embedder": embedder_name,
-                "duration": 0,
-                "file_size": 0,
-                "md5": "",
-                "embeddings": {},
-                "media_bytes": zf.read(member),
-                "media_string": None,
-                "filename": fname,
-                "category": ordered[0],
-                "categories": ordered,
-                "regions": regions,
-                "origin": {"importer": "staged_coco_val", "params": {"embedder": embedder_name}},
-                "origin_name": filenames[image_id],
-            }
-    if missing:
-        log(f"  coco: WARNING {missing} annotated images absent from the zip")
-
-
-# --------------------------------------------------------------------------
-# Box-size-banded Visual Genome, assembled from the full source
-# --------------------------------------------------------------------------
-
-
-#: The only per-category fields :func:`_band_categories` reads. Named once so
-#: the tolerance check below stays honest: if the selector ever starts reading
-#: a field the pre-envelope scans do not carry, this list is what makes the old
-#: shape fail loudly instead of selecting from partial statistics.
-_SCAN_FIELDS = ("voted_area", "n_images", "union_inflation")
-
-
-def load_box_scan_categories() -> dict[str, dict]:
-    """Read the full-VG box scan, tolerating both eras of its file format.
-
-    Two shapes exist in the wild, and the reader accepts either:
-
-    * **Pre-2026-08-17** — the bare ``{category: stats}`` dict at top level.
-      This is what the published ``vg_box_*`` sets were selected from.
-    * **2026-08-17 and later** — ``fb4f4ec03`` wrapped that dict in
-      ``{"meta": {...}, "categories": {...}}`` so per-band supply could be
-      recorded alongside it.
-
-    Tolerating both is deliberate rather than lazy. The envelope was the *only*
-    incompatibility: the newer scan also carries ``bands``, ``bands_compact``,
-    ``n_compact`` and ``compact_frac`` per category, and
-    :func:`_band_categories` reads none of them -- only ``voted_area``,
-    ``n_images`` and ``union_inflation``, all three present in the old shape.
-    So an old scan still selects exactly what it always selected, and the
-    alternative repair (re-running ``scan_vg_boxes.py``) would *not*: the
-    current scanner applies per-image compact filtering (``10239c24e``) and
-    per-band supply (``fb4f4ec03``), which qualify categories differently and
-    would silently redefine three datasets whose numbers are published in
-    #3129 and #3156. Old scans also sit on other people's scratch, so the
-    reader is the right place to absorb this rather than the file.
-    """
-    scan_path = pc.PILE / "vg_box_scale.json"
-    if not scan_path.exists():
-        raise SystemExit(f"missing {scan_path}; run scan_vg_boxes.py first")
-    scan = json.loads(scan_path.read_text())
-    if not isinstance(scan, dict) or not scan:
-        raise SystemExit(f"{scan_path} is not a non-empty JSON object; re-run scan_vg_boxes.py")
-    # Discriminate on shape, not just on the key being present: a bare scan is
-    # keyed by VG's free-text vocabulary, so "categories" is a name it could in
-    # principle hold. An envelope's "categories" maps to the stats dict; a
-    # category of that name would map to a stats entry carrying `voted_area`.
-    envelope = isinstance(scan.get("categories"), dict) and "voted_area" not in scan["categories"]
-    stats = scan["categories"] if envelope else scan
-    if not isinstance(stats, dict) or not stats:
-        raise SystemExit(f"{scan_path} holds no categories; re-run scan_vg_boxes.py")
-    # Fail here rather than with a bare KeyError deep inside the comprehension:
-    # a scan missing a field the selector needs is a format problem, and saying
-    # so by name is what tells the next person which era the file is from.
-    missing = sorted({f for f in _SCAN_FIELDS for s in stats.values() if isinstance(s, dict) and f not in s})
-    if missing:
-        raise SystemExit(f"{scan_path} categories lack {', '.join(missing)}; re-run scan_vg_boxes.py")
-    return stats
-
-
-def _band_categories(band: str) -> list[str]:
-    """Pick this band's categories from the full-VG scan, stratified within it.
-
-    Stratified on purpose: taking the N best-supported categories in a band
-    would cluster them at one end of it (support correlates with size), so the
-    "band" would silently be a point. Splitting the band into N slots by
-    voted-area rank and taking the best-supported category in each keeps the
-    band spanning its own range.
-    """
-    stats = load_box_scan_categories()
-    lo, hi = pc.BOX_BANDS[band]
-
-    pool = [
-        (s["voted_area"], name)
-        for name, s in stats.items()
-        if lo <= s["voted_area"] < hi
-        and s["n_images"] >= pc.BAND_MIN_IMAGES
-        and s["union_inflation"] <= pc.BAND_MAX_INFLATION
-        and pc.is_object_category(name)
-    ]
-    if not pool:
-        raise SystemExit(f"no categories qualify for band {band!r}")
-    pool.sort()
-
-    if len(pool) < pc.BAND_N_CATEGORIES:
-        # Say so rather than quietly returning a shorter list: a band that
-        # cannot fill its quota is a real limit on what it can support.
-        log(
-            f"  band {band}: ONLY {len(pool)} categories qualify "
-            f"(wanted {pc.BAND_N_CATEGORIES}) -- band is supply-limited"
-        )
-    n = min(pc.BAND_N_CATEGORIES, len(pool))
-    chosen: list[str] = []
-    for i in range(n):
-        slot = pool[i * len(pool) // n : max((i + 1) * len(pool) // n, i * len(pool) // n + 1)]
-        best = max(slot, key=lambda t: stats[t[1]]["n_images"])
-        chosen.append(best[1])
-    log(f"  band {band}: {len(chosen)} categories from {len(pool)} candidates")
-    return sorted(set(chosen))
-
-
-def _load_vg_band(band: str, medias: dict[int, dict], embedder_name: str) -> None:
-    """Populate *medias* with full-VG images carrying this band's categories."""
-    import random  # noqa: PLC0415
-
-    from PIL import Image  # noqa: PLC0415
-
-    vg_root = pc.DEMO_CACHE / "visual_genome"
-    objects_json = vg_root / "objects.json"
-    if not objects_json.exists():
-        raise SystemExit(f"missing {objects_json}")
-
-    wanted = set(_band_categories(band))
-    paths: dict[int, Path] = {}
-    for d in (vg_root / "VG_100K", vg_root / "VG_100K_2"):
-        for p in d.iterdir():
-            if p.suffix.lower() == ".jpg":
-                try:
-                    paths[int(p.stem)] = p
-                except ValueError:
-                    continue
-
-    with objects_json.open() as fh:
-        records = json.load(fh)
-
-    # Every image carrying at least one of the band's categories.
-    hits: list[tuple[int, dict]] = []
-    for rec in records:
-        iid = int(rec["image_id"])
-        if iid not in paths:
-            continue
-        by_name: dict[str, list[list[float]]] = defaultdict(list)
-        for obj in rec.get("objects") or []:
-            names = obj.get("names") or []
-            if not names:
-                continue
-            name = str(names[0]).strip().lower()
-            if name not in wanted:
-                continue
-            x, y = float(obj.get("x", 0)), float(obj.get("y", 0))
-            w, h = float(obj.get("w", 0)), float(obj.get("h", 0))
-            if w > 0 and h > 0:
-                by_name[name].append([x, y, x + w, y + h])
-        if by_name:
-            hits.append((iid, dict(by_name)))
-
-    rng = random.Random(0xB0FFED)  # deterministic sample, stable across rebuilds
-    rng.shuffle(hits)
-    hits = hits[: pc.BAND_MAX_IMAGES]
-    log(f"  band {band}: {len(hits)} images carry a band category")
-
-    for iid, by_name in hits:
-        path = paths[iid]
-        try:
-            with Image.open(path) as im:
-                W, H = im.size
-            data = path.read_bytes()
-        except Exception:  # noqa: BLE001 - a corrupt file just drops out
-            continue
-        if W <= 0 or H <= 0:
-            continue
-        regions = [
-            {"box": [b[0] / W, b[1] / H, b[2] / W, b[3] / H], "label": name}
-            for name, boxes in by_name.items()
-            for b in boxes
-        ]
-        counts = {name: len(b) for name, b in by_name.items()}
-        ordered = [c for c, _ in sorted(counts.items(), key=lambda kv: (-kv[1], kv[0]))]
-        medias[iid] = {
-            "id": iid,
-            "media_type": "image",
-            "embedder": embedder_name,
-            "duration": 0,
-            "file_size": 0,
-            "md5": "",
-            "embeddings": {},
-            "media_bytes": data,
-            "media_string": None,
-            "filename": path.name,
-            "category": ordered[0],
-            "categories": ordered,
-            "regions": regions,
-            "origin": {"importer": "vg_box_band", "params": {"band": band, "embedder": embedder_name}},
-            "origin_name": str(path),
-        }
-
-
-def _vg_image_paths() -> dict[int, Path]:
-    """``{image_id: path}`` over both VG image dirs."""
-    vg_root = pc.DEMO_CACHE / "visual_genome"
-    paths: dict[int, Path] = {}
-    for d in (vg_root / "VG_100K", vg_root / "VG_100K_2"):
-        for p in d.iterdir():
-            if p.suffix.lower() == ".jpg":
-                try:
-                    paths[int(p.stem)] = p
-                except ValueError:
-                    continue
-    return paths
-
-
-def _vg_source() -> tuple[dict[int, Path], list, dict[int, tuple[int, int]]]:
-    """``(image paths, objects.json records, image dims)`` for the whole VG source.
-
-    Dims come from ``scan_vg_boxes.py``'s cache when it exists (it always does
-    in practice -- the scan is what chooses the classes), and are read from the
-    JPEG headers otherwise, which costs ~30 s.
-    """
-    from concurrent.futures import ThreadPoolExecutor  # noqa: PLC0415
-
-    from PIL import Image  # noqa: PLC0415
-
-    objects_json = pc.DEMO_CACHE / "visual_genome" / "objects.json"
-    if not objects_json.exists():
-        raise SystemExit(f"missing {objects_json}")
-
-    paths = _vg_image_paths()
-    cache = pc.PILE / "vg_image_dims.json"
-    dims: dict[int, tuple[int, int]] = {}
-    if cache.exists():
-        raw = json.loads(cache.read_text())
-        # Unreadable images are cached as null, so a complete cache has one
-        # entry per file (see scan_vg_boxes._read_dims).
-        if len(raw) >= len(paths):
-            dims = {int(k): tuple(v) for k, v in raw.items() if v}  # type: ignore[misc]
-    if not dims:
-        log("  no dims cache; reading JPEG headers")
-
-        def one(item):
-            iid, path = item
-            try:
-                with Image.open(path) as im:
-                    return iid, im.size
-            except Exception:  # noqa: BLE001 - a corrupt file just drops out
-                return iid, None
-
-        with ThreadPoolExecutor(max_workers=16) as ex:
-            for iid, size in ex.map(one, paths.items(), chunksize=256):
-                if size:
-                    dims[iid] = size
-
-    with objects_json.open() as fh:
-        records = json.load(fh)
-    return paths, records, dims
-
-
-def _load_corrections() -> dict[tuple[int, str], dict]:
-    """``{(image_id, class): verdict}`` from the corrections file, if any.
-
-    Verdicts, not corrections: a row exists for every reviewed ``(image, class)``
-    pair whether or not the human disagreed, so review *coverage* is knowable.
-    Without that, "no bus here" and "nobody looked" are the same absence, and
-    every rate computed afterwards is biased by an unknown amount.
-
-    Written by ``ingest_slate.py``; absent until the first review lands, which
-    is why this returns empty rather than failing.
-
-    **Boxes here are NORMALISED, unlike VG's and COCO's** -- they come from the
-    app's ``region_box``. The space is validated on the way in and converted to
-    pixels once, by :func:`_correction_boxes_px`, so that everything downstream
-    of this function is in one space. See ``pile_config.CORRECTION_BOX_SPACE``.
-    """
-    path = Path(os.environ.get("VTS_CORRECTIONS", pc.PILE / "corrections.json"))
-    if not path.exists():
-        return {}
-    rows = json.loads(path.read_text())
-    out: dict[tuple[int, str], dict] = {}
-    for r in rows:
-        _assert_correction_box_space(r, path)
-        out[(int(r["image_id"]), r["class"])] = r
-    return out
-
-
-def _assert_correction_box_space(row: dict, path: Path) -> None:
-    """Refuse a correction row whose boxes are not in the declared space.
-
-    The space is a *declaration*, not a guess: a pixel-space box handed to the
-    normalised path is undetectable once it has been divided by (W, H) -- the
-    result is a small, plausible, entirely wrong box, which is #3281 in the
-    other direction. Checking it here costs one comparison per row and is the
-    only point at which the two spaces are still distinguishable.
-    """
-    space = row.get("box_space", pc.CORRECTION_BOX_SPACE)
-    where = f"{path.name}: image {row.get('image_id')} / {row.get('class')!r}"
-    if space != pc.CORRECTION_BOX_SPACE:
-        raise SystemExit(f"{where}: box_space {space!r}, expected {pc.CORRECTION_BOX_SPACE!r}")
-    for b in row.get("boxes") or []:
-        if len(b) != 4:
-            raise SystemExit(f"{where}: box {b} is not [x0, y0, x1, y1]")
-        if not all(-1e-6 <= float(v) <= 1.0 + 1e-6 for v in b):
-            raise SystemExit(
-                f"{where}: box {b} has a coordinate outside [0, 1], so it is in PIXEL space "
-                f"while the file declares {pc.CORRECTION_BOX_SPACE!r}"
-            )
-        if float(b[2]) <= float(b[0]) or float(b[3]) <= float(b[1]):
-            raise SystemExit(f"{where}: box {b} is degenerate (x1 <= x0 or y1 <= y0)")
-
-
-def _correction_boxes_px(row: dict, W: int, H: int) -> list[list[float]]:
-    """A verdict's boxes in the pixel space of ``(W, H)``.
-
-    ``(W, H)`` is the space the image's *other* boxes were measured in -- the
-    COCO original for an anchored image, the VG copy otherwise -- because that
-    is what the region write later divides by. Scaling up here and dividing down
-    there is an exact round trip, so the stored box is the reviewer's box to the
-    last bit rather than merely close to it.
-    """
-    return [[float(b[0]) * W, float(b[1]) * H, float(b[2]) * W, float(b[3]) * H] for b in (row.get("boxes") or [])]
-
-
-def scale_label_digest(medias: dict[int, dict]) -> str:
-    """A hash of exactly what ``vg_scale_any`` copies out of ``vg_scale``.
-
-    ``vg_scale_any`` is a *relabel* of the built ``vg_scale`` pickle, so a fix to
-    ``vg_scale``'s labels, boxes or bands leaves the derived cell holding the old
-    ones -- with the right media count, the right vectors and a healthy-looking
-    ``--verify``. #3281 is the case: the box repair moves 97 images between
-    bands, and ``build_pile.py --force vg_scale`` alone would ship a
-    ``vg_scale_any`` still carrying the pre-repair regions.
-
-    Vectors are deliberately not in it: they are identical by construction (the
-    derived build never re-embeds) and ``cell_fingerprint`` already covers them.
-    What this pins is the half a rebuild of the parent can actually change.
-    """
-    h = hashlib.sha256()
-    for mid in sorted(medias):
-        m = medias[mid]
-        h.update(
-            json.dumps(
-                [
-                    mid,
-                    m.get("category"),
-                    m.get("categories"),
-                    m.get("evaluable_categories"),
-                    [
-                        [r.get("label"), [round(float(v), 9) for v in r.get("box") or []]]
-                        for r in m.get("regions") or []
-                    ],
-                ],
-                sort_keys=True,
-            ).encode()
-        )
-    return h.hexdigest()
-
-
-def region_geometry_problems(medias: dict[int, dict]) -> list[str]:
-    """Geometry no honest normalised region box can have (#3281).
-
-    The band check in :func:`verify` cannot see a coordinate-space mistake made
-    *before* banding, because the band is computed from the very box it would be
-    checking: crush a box to the origin and it is filed under ``@small``, where
-    a sub-pixel area is exactly what the band's name claims. Both sides move
-    together and the cell stays self-consistent. So the box has to be checked
-    against the frame rather than against its own label.
-
-    Two rules, and they are different in kind:
-
-    * **Sub-pixel** is absolute. A side below ``MIN_BOX_SIDE`` is under one pixel
-      on any image the pile holds, so no such box was ever drawn or annotated.
-      One is a failure.
-    * **Crushed to the origin** is a rate. A real small object can sit in the
-      top-left corner and 1.2% of healthy boxes do, so a single hit proves
-      nothing; a *population* of them is a double-normalise, which put 100% of
-      the affected images there.
-    """
-    problems: list[str] = []
-    n_boxes = 0
-    subpixel: list[str] = []
-    cornered = 0
-    edge = pc.CORNER_AREA_FRAC**0.5
-    for mid, m in medias.items():
-        for r in m.get("regions") or []:
-            b = r.get("box") or []
-            if len(b) != 4:
-                problems.append(f"media {mid} / {r.get('label')!r}: box {b} is not [x0, y0, x1, y1]")
-                continue
-            n_boxes += 1
-            if (b[2] - b[0]) < pc.MIN_BOX_SIDE or (b[3] - b[1]) < pc.MIN_BOX_SIDE:
-                subpixel.append(f"media {mid} / {r.get('label')!r} {[round(v, 6) for v in b]}")
-            if b[2] <= edge and b[3] <= edge:
-                cornered += 1
-    if subpixel:
-        problems.append(
-            f"{len(subpixel)} region boxes are sub-pixel (side < {pc.MIN_BOX_SIDE:g} of the frame), "
-            f"which no drawn or annotated box is -- e.g. {'; '.join(subpixel[:3])}"
-        )
-    if n_boxes and cornered / n_boxes > pc.MAX_CORNER_RATE:
-        problems.append(
-            f"{cornered}/{n_boxes} ({cornered / n_boxes:.1%}) of region boxes lie wholly inside the "
-            f"top-left {pc.CORNER_AREA_FRAC:.0%} of the frame, against a healthy rate near 1% -- "
-            f"the signature of a box normalised twice"
-        )
-    return problems
-
-
-def _load_vg_scale(medias: dict[int, dict], embedder_name: str) -> None:
-    """One pickle holding every ``(class, band)`` cell of the scale study (#3156).
-
-    The construction, in one paragraph: one image pool and one class list
-    (:data:`pile_config.SCALE_CLASSES`); for a class *c* and band *B* an image is
-    a **positive** when its compact union box for *c* falls in *B*, a
-    **negative** when it holds no instance of any class in *C*, and **excluded**
-    otherwise -- it holds *c* at some other size, so scoring it as a negative
-    would penalise a detector for finding a real bus, which is what #3156 is
-    about. Exclusion is carried per media as ``evaluable_categories`` and
-    honoured by ``vtscore.eval.labels.evaluable_pool``.
-
-    Cells are *designated* rather than inferred: exactly ``SCALE_N_POS``
-    positives and one shared pool of ``SCALE_N_NEG`` negatives each. Every cell
-    therefore has identical prevalence and identical negatives, so a
-    small-vs-large difference is a paired contrast on one class rather than two
-    datasets of different difficulty.
-
-    **The labels are COCO's, and the pool is the half of VG that can carry
-    them.** VG's own annotation is not exhaustive and measurably fails this
-    construction -- see the note in the body and ``coco_anchor.py``.
-    """
-    from PIL import Image  # noqa: PLC0415
-
-    import coco_anchor as ca  # noqa: PLC0415
-
-    wanted = set(pc.SCALE_CLASSES)
-    bands = list(pc.BOX_BANDS)
-    cells = [pc.scale_cell(c, b) for c in pc.SCALE_CLASSES for b in bands]
-
-    paths = _vg_image_paths()
-    _, records, dims = _vg_source()
-
-    # --- labels: VG, repaired where a better reference exists ---------------
-    #
-    # VG's own annotation is not exhaustive: measured against COCO its recall
-    # over C is 0.61, and 1.35% of the images it treats as negatives actually
-    # hold the object -- ~54 hidden positives against 100 labelled ones per cell,
-    # and 4.1% for `backpack` puts more real backpacks among the negatives than
-    # among the positives (`coco_anchor.py`, issue #3156).
-    #
-    # 48% of VG's images ARE COCO images, and COCO annotates C exhaustively, so
-    # on that half the repair is free and total. The other half has no reference
-    # and is what the human slates are for (`make_audit_slate.py`); its verdicts
-    # arrive here through the same corrections file. The pool stays the whole of
-    # VG either way -- restricting it to the COCO half would make this a COCO
-    # subset with extra steps, losing VG's non-COCO diversity for nothing.
-    image_data, instances = ca.ensure_sources(pc.PILE / "coco_anchor", fetch=False)
-    truth = ca.coco_truth(instances, wanted)
-    with image_data.open() as fh:
-        coco_of = {int(m["image_id"]): int(m["coco_id"]) for m in json.load(fh) if m.get("coco_id")}
-
-    corrections = _load_corrections()
-    log(f"  {len(coco_of)} VG images carry a coco_id; {len(corrections)} human verdicts on file")
-
-    # image -> {class: [boxes]}, and whether the image's labels are exhaustive.
-    labels: dict[int, dict[str, list[list[float]]]] = {}
-    exhaustive: set[int] = set()
-    for rec in records:
-        iid = int(rec["image_id"])
-        if iid not in paths or dims.get(iid) is None:
-            continue
-        by_name: dict[str, list[list[float]]] = defaultdict(list)
-        for obj in rec.get("objects") or []:
-            names = obj.get("names") or []
-            if not names:
-                continue
-            name = str(names[0]).strip().lower()
-            if name not in wanted:
-                continue
-            x, y = float(obj.get("x", 0)), float(obj.get("y", 0))
-            w, h = float(obj.get("w", 0)), float(obj.get("h", 0))
-            if w > 0 and h > 0:
-                by_name[name].append([x, y, x + w, y + h])
-        labels[iid] = dict(by_name)
-
-    # The pixel space each image's boxes live in. VG ships DOWNSCALED copies of
-    # the COCO originals -- 500 px wide against COCO's 640, on 95% of the
-    # overlap -- so a COCO box normalised by the VG file's dimensions lands in
-    # the wrong place and with the wrong extent. Normalise every box by the
-    # dimensions of the image its coordinates were measured on.
-    box_dims: dict[int, tuple[int, int]] = dict(dims)
-    n_anchored = 0
-    n_reframed = 0
-    for iid in labels:
-        cid = coco_of.get(iid)
-        ref = truth.get(cid) if cid is not None else None
-        if ref is None:
-            continue
-        wh = ca.COCO_DIMS.get(cid)
-        if wh is None:
-            continue
-        # A normalised box only transfers between two copies of an image if they
-        # frame the same thing. 49 of the 51,497 overlaps disagree on aspect
-        # ratio -- some are transposed (VG 500x375 against COCO 375x500), i.e. a
-        # rotated or re-cropped copy -- and there COCO's box does not describe
-        # VG's pixels at all. Those keep VG's own labels and stay unanchored
-        # rather than importing a box that points at the wrong part of a
-        # different framing.
-        vw, vh = dims[iid]
-        if abs((vw / vh) - (wh[0] / wh[1])) / (wh[0] / wh[1]) > pc.MAX_ASPECT_DRIFT:
-            n_reframed += 1
-            continue
-        box_dims[iid] = wh
-        # COCO's annotation REPLACES VG's for this image rather than merging
-        # with it: the two disagree in both directions, and only one of them is
-        # exhaustive. Keeping VG's extra boxes would reintroduce exactly the
-        # unverifiable labels this is repairing.
-        labels[iid] = {name: bs for name, bs in ref.items() if name in wanted}
-        exhaustive.add(iid)
-        n_anchored += 1
-
-    # A pair a reviewer ruled "present" without drawing a box cannot be banded:
-    # a band is a claim about size, and no size was measured. It leaves every
-    # cell of that class instead -- neither a positive nor a negative -- which
-    # is precisely what the third value is for.
-    #
-    # The box, when there is one, is NORMALISED and everything in `labels` is in
-    # pixels, so it is converted here -- once, against the same (W, H) the region
-    # write divides by. Merging it unconverted is #3281: the region write then
-    # normalises a normalised coordinate, which divides it by ~500 and parks the
-    # box on the frame origin. Nothing downstream could see that, because the
-    # BAND is derived from the same corrupted box, so the cell name and its boxes
-    # stayed consistent with each other all the way into the study.
-    unbanded: set[tuple[int, str]] = set()
-    for (iid, name), verdict in corrections.items():
-        if iid not in labels:
-            continue
-        if verdict.get("present"):
-            boxes = _correction_boxes_px(verdict, *box_dims[iid])
-            if boxes:
-                labels[iid][name] = boxes
-            else:
-                labels[iid].pop(name, None)
-                unbanded.add((iid, name))
-        else:
-            labels[iid].pop(name, None)
-        exhaustive.add(iid)  # someone looked at this pair
-
-    log(
-        f"  labels: {len(labels)} VG images, {n_anchored} repaired from COCO, "
-        f"{len(exhaustive)} with a verified pair, {n_reframed} skipped as re-framed copies"
-    )
-
-    # --- candidates ---------------------------------------------------------
-    supply: dict[str, dict[str, list[int]]] = {c: {b: [] for b in bands} for c in pc.SCALE_CLASSES}
-    boxes_for: dict[tuple[int, str], list[list[float]]] = {}
-    clean: list[int] = []
-
-    for iid, by_name in labels.items():
-        W, H = box_dims[iid]
-        area = float(W * H)
-        if not by_name:
-            # Only a true negative for every class in C may join the shared pool.
-            if not any((iid, c) in unbanded for c in pc.SCALE_CLASSES):
-                clean.append(iid)
-            continue
-        for name, bs in by_name.items():
-            ux0 = min(b[0] for b in bs)
-            uy0 = min(b[1] for b in bs)
-            ux1 = max(b[2] for b in bs)
-            uy1 = max(b[3] for b in bs)
-            union = max(0.0, ux1 - ux0) * max(0.0, uy1 - uy0) / area
-            largest = max((b[2] - b[0]) * (b[3] - b[1]) for b in bs) / area
-            # Scattered instances in *this* image: the union box describes the
-            # scatter rather than the object, so the image is excluded from
-            # every band of this class rather than banded by a box no user
-            # would drag.
-            if union > largest * pc.BAND_MAX_INFLATION:
-                continue
-            for band, (lo, hi) in pc.BOX_BANDS.items():
-                if lo <= union < hi:
-                    supply[name][band].append(iid)
-                    boxes_for[(iid, pc.scale_cell(name, band))] = bs
-                    break
-
-    # Selection must be stable under a changing candidate list, not merely
-    # deterministic. `rng.sample` is deterministic given the same list, but any
-    # edit to the pool -- a label fix, an image excluded as a re-framed copy --
-    # reshuffles the entire draw, and a rebuild then silently retires images a
-    # human already reviewed (49 of 360 in one such rebuild). Ranking each
-    # candidate by a hash of (cell, image_id) instead means adding or removing
-    # one image changes only that image's membership.
-    def _rank(cell: str, iid: int) -> str:
-        return hashlib.sha1(f"{cell}:{iid}".encode()).hexdigest()  # noqa: S324 - not security
-
-    # A roster pins the membership a review was actually carried out against.
-    # Without it, switching selection rules retires images a human has already
-    # judged -- the hash draw and the earlier random draw share only ~228 of
-    # 3,900 negatives, so the entire negative review would have been orphaned.
-    # Entries that are no longer eligible (a correction moved or removed them)
-    # drop out and the shortfall is backfilled by rank, so the roster adapts
-    # without ever reshuffling what it can keep.
-    roster = {}
-    if pc.ROSTER.exists():
-        roster = json.loads(pc.ROSTER.read_text())
-        log(f"  roster: {pc.ROSTER.name} pins {len(roster.get('cells', {}))} cells")
-
-    chosen: dict[str, list[int]] = {}
-    for c in pc.SCALE_CLASSES:
-        for band in bands:
-            pool = sorted(supply[c][band])
-            cell = pc.scale_cell(c, band)
-            if len(pool) < pc.SCALE_N_POS:
-                # Say so rather than quietly building a smaller cell: unequal
-                # prevalence between bands is the defect this construction
-                # exists to remove.
-                log(f"  UNDER-SUPPLIED {cell}: {len(pool)} positives (wanted {pc.SCALE_N_POS})")
-            eligible = set(pool)
-            pinned = [i for i in roster.get("cells", {}).get(cell, []) if i in eligible]
-            # A correction can move an image to another band -- that is the
-            # point of re-drawing a box. If the destination cell is already full
-            # of images nobody has looked at, the reviewed one lands nowhere and
-            # the review quietly stops covering it (99 of 360 boxed positives,
-            # first time round). Reviewed images therefore outrank unreviewed
-            # ones for a seat, wherever their box now puts them.
-            reviewed = {i for i in eligible if (i, c) in corrections}
-            order = (
-                [i for i in pinned if i in reviewed]
-                + sorted(reviewed - set(pinned), key=lambda i: _rank(cell, i))
-                + [i for i in pinned if i not in reviewed]
-                + sorted(eligible - reviewed - set(pinned), key=lambda i: _rank(cell, i))
-            )
-            chosen[cell] = order[: pc.SCALE_N_POS]
-
-    clean.sort()
-    # Draw spares beyond the designated pool. A human verdict can retire a
-    # contaminated negative later, and re-designating from spares costs a
-    # relabel rather than a re-embed of every cell.
-    want_neg = pc.SCALE_N_NEG + pc.SCALE_N_NEG_SPARE
-    clean_set = set(clean)
-    drawn = [i for i in roster.get("negatives", []) + roster.get("spares", []) if i in clean_set]
-    if len(drawn) < want_neg:
-        extra = sorted(clean_set - set(drawn), key=lambda i: _rank("__negatives__", i))
-        drawn += extra[: want_neg - len(drawn)]
-    drawn = drawn[:want_neg]
-    negatives, spares = drawn[: pc.SCALE_N_NEG], drawn[pc.SCALE_N_NEG :]
-    neg_set = set(negatives)
-    pc.ROSTER.write_text(json.dumps({"cells": chosen, "negatives": negatives, "spares": spares}, indent=1) + "\n")
-    log(
-        f"  {sum(len(v) for v in chosen.values())} positives over {len(cells)} cells, "
-        f"{len(negatives)} shared negatives + {len(spares)} spares (from {len(clean)} clean images)"
-    )
-
-    # media id -> the cells it is a positive for. Negatives get every cell.
-    positive_in: dict[int, list[str]] = defaultdict(list)
-    for cell, ids in chosen.items():
-        for iid in ids:
-            positive_in[iid].append(cell)
-
-    for iid in sorted(set(positive_in) | set(negatives) | set(spares)):
-        path = paths[iid]
-        try:
-            with Image.open(path) as im:
-                vw, vh = im.size
-            data = path.read_bytes()
-        except Exception:  # noqa: BLE001 - a corrupt file just drops out
-            continue
-        if vw <= 0 or vh <= 0:
-            continue
-        # Normalised region boxes are resolution-independent, so they must be
-        # divided by the size of the image the coordinates came from -- which is
-        # the COCO original for a repaired image, not the VG copy carrying the
-        # pixels.
-        W, H = box_dims[iid]
-        cats = sorted(positive_in.get(iid, []))
-        regions = [
-            {"box": [b[0] / W, b[1] / H, b[2] / W, b[3] / H], "label": cell}
-            for cell in cats
-            for b in boxes_for.get((iid, cell), [])
-        ]
-        medias[iid] = {
-            "id": iid,
-            "media_type": "image",
-            "embedder": embedder_name,
-            "duration": 0,
-            "file_size": 0,
-            "md5": "",
-            "embeddings": {},
-            "media_bytes": data,
-            "media_string": None,
-            "filename": path.name,
-            "category": cats[0] if cats else "",
-            "categories": cats,
-            # A designated cell membership, not a closed world: a positive is
-            # scorable only in the cells it was drawn for, and the shared
-            # negatives are scorable everywhere.
-            "evaluable_categories": cats if cats else (list(cells) if iid in neg_set else []),
-            # Whether this image's labels rest on an exhaustive reference (COCO,
-            # or a human who looked). False means VG's silence is the only
-            # evidence of absence -- which is what the review slates target.
-            "labels_exhaustive": iid in exhaustive,
-            "regions": regions,
-            "origin": {"importer": "vg_scale", "params": {"embedder": embedder_name, "labels": "coco"}},
-            "origin_name": str(path),
-        }
-
-    # Refuse to embed a pickle whose boxes are impossible. `--verify` runs the
-    # same check, but only after the GPU hours are spent and the cell is on
-    # disk; #3281 got as far as three published studies that way.
-    bad = region_geometry_problems(medias)
-    if bad:
-        raise SystemExit("vg_scale: " + "; ".join(bad))
-
-
-def _load_vg_scale_any(medias: dict[int, dict], embedder_name: str) -> None:
-    """``vg_scale`` with the box-size band collapsed away (#3115).
-
-    #3156 built ``vg_scale`` to ask a question about *scale*, so its cells are
-    keyed ``class@band`` and each holds exactly 100 positives.  A study that does
-    not care how big the box is wants the same verified images with the band
-    dropped: 12 classes, **300 positives each**, one shared negative pool, and
-    identical prevalence everywhere.  That is what the calibration studies
-    actually need, and what ``visual_genome_m`` conspicuously is not - its
-    selected categories run from 25 positives (``banana``) to 1645
-    (``building``), and the thin ones produce cells with *no trainable step at
-    all* (``ball``, 51 positives, wrote a header and zero rows).
-
-    Derived from the built ``vg_scale`` pickle rather than re-scanned from the
-    VG source, so it is **the same images, the same boxes and the same
-    hand-checked label corrections** with one string operation applied - and it
-    costs no image decode and no embedding pass.
-
-    The whole build is a relabel, ``class@band -> class``, applied to
-    ``category`` / ``categories`` / ``evaluable_categories`` and to each region's
-    ``label``.  What matters is what that does *not* change, because the
-    exclusion semantics are the point of #3156 and the easy version of this
-    build destroys them:
-
-    * The **300 medias evaluable on nothing** stay evaluable on nothing.  They
-      hold one of the 12 classes at a size outside every band, so scoring them
-      as negatives would penalise a detector for finding a real bus - exactly
-      what #3156 exists to prevent.  A naive "positive if in any band, negative
-      otherwise" rule turns all 300 into negatives.
-    * A media positive for one class stays evaluable **only** for that class.
-      VG's labels are not exhaustive (``labels_exhaustive: False``), so an image
-      of a dog is not evidence of the absence of a clock.
-    * The 3900-image clean pool - images holding no instance of any of the 12 -
-      stays evaluable for all of them.
-
-    Result per class: 300 positives against 3900 shared negatives, i.e. a 4200
-    -image evaluable pool at **7.1 % prevalence, identical for all twelve**.
-    """
-    src = pc.cell_path("vg_scale", embedder_name)
-    if not src.exists():
-        raise SystemExit(f"vg_scale_any is derived from {src.name}, which does not exist yet - build vg_scale first")
-
-    def _base(label: str) -> str:
-        return label.split("@", 1)[0]
-
-    def _collapse(labels) -> list[str]:
-        """Base classes, order-preserving and deduped."""
-        return list(dict.fromkeys(_base(c) for c in (labels or [])))
-
-    loaded = _cells_io().load_medias(src)
-    parent_digest = scale_label_digest(loaded)
-    log(f"  derived from {src.name}: {len(loaded)} medias, parent labels {parent_digest[:12]}")
-    for mid, media in loaded.items():
-        d = dict(media)
-        d["categories"] = _collapse(d.get("categories"))
-        # `evaluable_categories` is collapsed rather than rebuilt: an empty list
-        # must survive as an empty list (see the docstring), and rebuilding it
-        # from the class roster is precisely the bug that would not.
-        d["evaluable_categories"] = _collapse(d.get("evaluable_categories"))
-        if d.get("category"):
-            d["category"] = _base(d["category"])
-        d["regions"] = [{**r, "label": _base(r["label"])} for r in (d.get("regions") or [])]
-        # The parent's label digest travels with every media, so `--verify` can
-        # tell a derived cell that is merely older than its parent from one that
-        # no longer agrees with it. Without it a stale derivation is invisible.
-        d["origin"] = {
-            "importer": "vg_scale_any",
-            "params": {"embedder": embedder_name, "derived_from": src.name, "parent_labels": parent_digest},
-        }
-        medias[mid] = d
-
-    n_pos = sum(1 for d in medias.values() if d["categories"])
-    n_clean = sum(1 for d in medias.values() if not d["categories"] and d["evaluable_categories"])
-    n_excluded = sum(1 for d in medias.values() if not d["categories"] and not d["evaluable_categories"])
-    log(f"  {n_pos} positives, {n_clean} shared negatives, {n_excluded} excluded-everywhere (must be nonzero)")
-    if not n_excluded:
-        raise SystemExit("vg_scale_any: no excluded-everywhere medias survived - the exclusion semantics were lost")
-
-
-def _load_demo(dataset: str, medias: dict[int, dict], embedder_name: str) -> None:
-    from vtscore.datasets.loader_demo import load_demo_dataset  # noqa: PLC0415
-
-    load_demo_dataset(dataset, medias, embedder_name=embedder_name)
-
-
-# --------------------------------------------------------------------------
-# Build
-# --------------------------------------------------------------------------
+from pilebuild.audit import list_cells, rebuildable, report_bands, verify  # noqa: E402
+from pilebuild.boxscan import band_categories, load_box_scan_categories  # noqa: E402
+from pilebuild.env import assert_vtscore_is_this_checkout, cells_io, log  # noqa: E402
+from pilebuild.geometry import region_geometry_problems, scale_label_digest  # noqa: E402
+from pilebuild.loaders import loader_for  # noqa: E402
+from pilebuild.manifest import write_manifest  # noqa: E402
+from pilebuild.provenance import cell_fingerprint, write_provenance  # noqa: E402
+from pilebuild.provenance_report import provenance_report  # noqa: E402
+from pilebuild.vgsource import vg_image_paths  # noqa: E402
+
+#: Names sibling scripts and tests import off this module. Spelled out so that
+#: moving a definition into ``pilebuild/`` cannot quietly break an importer that
+#: this repo's test suite does not cover (``precision/build_arm.py``,
+#: ``fastproc/build_arm.py``, and the five ``pile/make_*.py`` sheet builders).
+__all__ = [
+    "assert_vtscore_is_this_checkout",
+    "band_categories",
+    "build_cell",
+    "cell_fingerprint",
+    "list_cells",
+    "load_box_scan_categories",
+    "main",
+    "provenance_report",
+    "rebuildable",
+    "region_geometry_problems",
+    "report_bands",
+    "scale_label_digest",
+    "verify",
+    "vg_image_paths",
+    "write_manifest",
+    "write_provenance",
+]
+
+#: Pre-split spellings, kept because out-of-tree callers and older notebooks use
+#: them. ``_band_categories`` in particular is what ``tests_lib`` calls today.
+_band_categories = band_categories
+_vg_image_paths = vg_image_paths
+_cells_io = cells_io
 
 
 @contextmanager
@@ -992,200 +116,6 @@ def _embed_batch_size(embedder: str):
         os.environ.pop("VTSEARCH_EMBED_BATCH_SIZE", None)
 
 
-# --------------------------------------------------------------------------
-# Provenance: which machine produced this cell (#3160)
-# --------------------------------------------------------------------------
-
-
-def _device_record() -> dict:
-    """Everything about the machine that a later reader needs to compare cells.
-
-    ``gres/gpu:v100`` is a *type*, and #3143 measured that a type is not a
-    device: two nodes both answering to it produced ``siglip2_l`` vectors 1.5e-04
-    apart, while three other devices agreed to ~1e-12. Nothing in ``scontrol`` or
-    ``--gres`` distinguishes the parts, so the only way a rebuild can be told
-    apart from the cell it replaces is if the build **writes down** what it ran
-    on. That is what this is; it does not make the arithmetic reproducible, it
-    makes the difference visible.
-    """
-    import torch  # noqa: PLC0415
-
-    from vtscore.config import EMBED_PRECISION, embed_precision  # noqa: PLC0415
-
-    rec: dict = {
-        "hostname": os.uname().nodename,
-        # The host, not the card, turned out to be the axis (#3160): PyTorch's
-        # CPU kernel dispatch decides how the 384px resize rounds, and an AVX2
-        # host disagrees with an AVX-512 one on 12.3% of pixels by one 8-bit
-        # level -- which is the whole of the 1.5e-04 that #3143 attributed to the
-        # GPU. Pinning the dispatch removes it; recording it explains a cell that
-        # was built before the pin.
-        "cpu": _cpu_model(),
-        # What the process RESOLVED, not what was asked for. `ATEN_CPU_CAPABILITY`
-        # is advisory: an unsupported or misspelled value is ignored in silence,
-        # so echoing the request would record a pin that never took. Both are
-        # written; a reader can see a request that did not land.
-        "cpu_capability": _cpu_capability(),
-        "aten_cpu_capability_requested": os.environ.get("ATEN_CPU_CAPABILITY"),
-        "slurm_job": os.environ.get("SLURM_JOB_ID"),
-        "slurm_gres": os.environ.get("SLURM_JOB_GRES") or os.environ.get("SBATCH_GRES"),
-        "precision_requested": EMBED_PRECISION,
-        "precision_resolved": embed_precision(),
-        "torch": torch.__version__,
-        "cuda_runtime": getattr(torch.version, "cuda", None),
-        # The node is not the only unrecorded axis. `requirements/image-embedders.txt`
-        # pins `transformers>=4.49`, and v5 renamed the image processors: the plain
-        # name is now the torchvision implementation and the PIL one moved to a
-        # `Pil` suffix. So two hosts resolving different versions preprocess the
-        # same image differently -- measured at 7.8e-3 max abs in pixels between
-        # the two paths, well above the 1.5e-04 device effect this record was
-        # written for. Recording the version and the class that actually loaded
-        # costs nothing and makes that axis visible too.
-        "transformers": _transformers_version(),
-        "commit": _git_commit(),
-    }
-    if not torch.cuda.is_available():
-        rec["gpu_name"] = None
-        rec["note"] = "no CUDA device; embedded on CPU"
-        return rec
-    props = torch.cuda.get_device_properties(0)
-    major, minor = torch.cuda.get_device_capability(0)
-    rec.update(
-        {
-            "gpu_name": torch.cuda.get_device_name(0),
-            "gpu_capability": f"sm_{major}{minor}",
-            # SM count was #3143's leading hypothesis for the drift (different
-            # tiling, different accumulation order). It is not the cause -- the
-            # two V100 parts have 80 SMs each and produce bit-identical GEMMs --
-            # but it stays in the record because it is the cheapest way to tell
-            # two cards apart that share a name.
-            "multi_processor_count": props.multi_processor_count,
-            "total_memory_gb": round(props.total_memory / 1e9, 1),
-            "cudnn_version": torch.backends.cudnn.version(),
-            "cudnn_allow_tf32": bool(torch.backends.cudnn.allow_tf32),
-            "matmul_allow_tf32": bool(torch.backends.cuda.matmul.allow_tf32),
-            "cudnn_deterministic": bool(torch.backends.cudnn.deterministic),
-            "cudnn_benchmark": bool(torch.backends.cudnn.benchmark),
-        }
-    )
-    return rec
-
-
-def _cpu_capability() -> str | None:
-    """The CPU kernel ISA torch is actually dispatching to (``AVX512``/``AVX2``/...)."""
-    import torch  # noqa: PLC0415
-
-    getter = getattr(getattr(torch.backends, "cpu", None), "get_cpu_capability", None)
-    return str(getter()) if getter else None
-
-
-def _cpu_model() -> str | None:
-    """The host CPU model string, or None where /proc/cpuinfo is not readable."""
-    try:
-        for line in Path("/proc/cpuinfo").read_text().splitlines():
-            if line.startswith("model name"):
-                return line.split(":", 1)[1].strip()
-    except OSError:
-        pass
-    return None
-
-
-def _transformers_version() -> str | None:
-    try:
-        import transformers  # noqa: PLC0415
-    except ImportError:
-        return None
-    return getattr(transformers, "__version__", None)
-
-
-def _processor_record(embedder: str) -> dict:
-    """The preprocessing classes this embedder actually resolved to.
-
-    Best effort: an embedder with no HF processor (or one that failed to load)
-    records nulls rather than sinking a build that has already produced a cell.
-    """
-    try:
-        from vtscore.media import get_embedder  # noqa: PLC0415
-
-        emb = get_embedder(embedder)
-        proc = getattr(emb, "_processor", None)
-        image_proc = getattr(proc, "image_processor", None)
-        return {
-            "processor_class": type(proc).__name__ if proc is not None else None,
-            "image_processor_class": type(image_proc).__name__ if image_proc is not None else None,
-        }
-    except Exception as exc:  # noqa: BLE001 -- provenance must never fail a build
-        return {"processor_class": None, "image_processor_class": None, "error": repr(exc)[:120]}
-
-
-def _git_commit() -> str | None:
-    """The commit of the checkout that is about to embed, or None outside git."""
-    import subprocess  # noqa: PLC0415, S404 -- fixed argv, no shell
-
-    repo = Path(os.environ.get("VTS_REPO") or Path(__file__).resolve().parents[3])
-    try:
-        out = subprocess.run(  # noqa: S603
-            ["git", "-C", str(repo), "rev-parse", "HEAD"],  # noqa: S607
-            capture_output=True,
-            text=True,
-            timeout=15,
-            check=False,
-        )
-    except (OSError, subprocess.SubprocessError):
-        return None
-    return out.stdout.strip() or None
-
-
-def cell_fingerprint(dataset: str, embedder: str, medias: dict | None = None) -> dict:
-    """A hash of the cell's vectors, in a fixed media-id order.
-
-    The point of the hash is that it survives the cell it describes: a rebuild
-    can be compared against it without keeping the old 900 MB pickle, which is
-    exactly the check a purge-and-rebuild needs and cannot otherwise make.
-
-    ``medias`` lets a fresh build hand over what it already has; the patch cells
-    are 3.5 GB on disk, and re-reading one purely to hash it would add minutes
-    and a second copy in RAM to every build.
-    """
-    import hashlib  # noqa: PLC0415
-
-    import numpy as np  # noqa: PLC0415
-
-    from vtscore.embedding.media_vectors import media_embedding  # noqa: PLC0415
-
-    if medias is None:
-        medias = _cells_io().load_medias(pc.cell_path(dataset, embedder))
-    ids = sorted(medias)
-    vecs = [media_embedding(medias[i]) for i in ids]
-    arr = np.stack([np.asarray(v, dtype=np.float32) for v in vecs if v is not None])
-    digest = hashlib.sha256(arr.tobytes()).hexdigest()
-    return {
-        "n_vectors": int(arr.shape[0]),
-        "dim": int(arr.shape[1]) if arr.ndim > 1 else None,
-        "vectors_sha256": digest,
-        "id_range": [int(ids[0]), int(ids[-1])] if ids else None,
-    }
-
-
-def write_provenance(dataset: str, embedder: str, summary: dict, medias: dict | None = None) -> Path:
-    """Write the per-cell provenance sidecar."""
-    record = {
-        "dataset": dataset,
-        "embedder": embedder,
-        "cell": pc.cell_path(dataset, embedder).name,
-        "built_at": time.strftime("%Y-%m-%dT%H:%M:%S%z"),
-        "device": _device_record(),
-        "preprocessing": _processor_record(embedder),
-        "cell_summary": {k: v for k, v in summary.items() if k != "status"},
-        "fingerprint": cell_fingerprint(dataset, embedder, medias),
-    }
-    path = pc.provenance_path(dataset, embedder)
-    path.write_text(json.dumps(record, indent=2) + "\n")
-    dev = record["device"]
-    log(f"  provenance: {dev.get('gpu_name')} on {dev.get('hostname')} -> {path.name}")
-    return path
-
-
 def build_cell(dataset: str, embedder: str, force: bool = False) -> dict:
     """Build one cell, returning a summary record."""
     out = pc.cell_path(dataset, embedder)
@@ -1202,17 +132,7 @@ def build_cell(dataset: str, embedder: str, force: bool = False) -> dict:
     t0 = time.time()
 
     medias: dict[int, dict] = {}
-    if kind == "coco":
-        _load_coco(medias, embedder)
-    elif kind == "vg_band":
-        _load_vg_band(pc.DATASETS[dataset]["band"], medias, embedder)
-    elif kind == "vg_scale":
-        _load_vg_scale(medias, embedder)
-    elif kind == "vg_scale_any":
-        _load_vg_scale_any(medias, embedder)
-    else:
-        pc.require_demo_source(dataset)
-        _load_demo(dataset, medias, embedder)
+    loader_for(dataset, kind).load(dataset, medias, embedder)
     log(f"  loaded {len(medias)} medias in {time.time() - t0:.0f}s")
 
     from vtscore.datasets.stages.embedding import embed_missing  # noqa: PLC0415
@@ -1223,7 +143,7 @@ def build_cell(dataset: str, embedder: str, force: bool = False) -> dict:
     embed_s = time.time() - t1
 
     n_patch = sum(1 for m in medias.values() if m.get("patch_grid") is not None)
-    nbytes = _cells_io().dump_medias(medias, out)
+    nbytes = cells_io().dump_medias(medias, out)
     total_s = time.time() - t0
     log(
         f"  wrote {out.name}: {nbytes / 1e6:.0f} MB, {len(medias)} medias, "
@@ -1241,571 +161,6 @@ def build_cell(dataset: str, embedder: str, force: bool = False) -> dict:
     }
     write_provenance(dataset, embedder, summary, medias)
     return summary
-
-
-# --------------------------------------------------------------------------
-# Verify + manifest
-# --------------------------------------------------------------------------
-
-
-def verify() -> int:
-    """Load every present cell and check it is usable. Returns an exit code."""
-    io = _cells_io()
-    problems: list[str] = []
-    rows = []
-    counts_by_dataset: dict[str, dict[str, int]] = defaultdict(dict)
-    for ds, emb in pc.cells():
-        path = pc.cell_path(ds, emb)
-        if not path.exists():
-            rows.append((ds, emb, "MISSING", "", "", ""))
-            continue
-        medias = io.load_medias(path)
-        n = len(medias)
-        counts_by_dataset[ds][emb] = n
-        n_patch = sum(1 for m in medias.values() if m.get("patch_grid") is not None)
-        first = next(iter(medias.values()), None)
-        dim = ""
-        if first is not None:
-            from vtscore.embedding.media_vectors import media_embedding  # noqa: PLC0415
-
-            vec = media_embedding(first)
-            dim = str(len(vec)) if vec is not None else "NO-VECTOR"
-        want_region = pc.region_capable(ds, emb)
-        state = "ok"
-        if n == 0:
-            state = "EMPTY"
-            problems.append(f"{ds} x {emb}: 0 medias")
-        elif dim in ("", "NO-VECTOR"):
-            state = "NO-VECTOR"
-            problems.append(f"{ds} x {emb}: medias carry no embedding")
-        elif want_region and n_patch < n:
-            state = "PATCH-GAP"
-            problems.append(f"{ds} x {emb}: region-capable but patch_grid on only {n_patch}/{n}")
-        elif not pc.is_patch_embedder(emb) and n_patch:
-            state = "UNEXPECTED-PATCH"
-            problems.append(f"{ds} x {emb}: single-vector embedder carries patch grids")
-        rows.append((ds, emb, state, str(n), f"{n_patch}/{n}", dim))
-
-    # A banded cell's NAME asserts the size of its boxes, so the stored box has
-    # to agree with it. That is what catches a coordinate-space mistake: VG
-    # ships 500 px copies of COCO's 640 px originals, and normalising a COCO box
-    # by the VG file's dimensions leaves every box shifted and mis-scaled while
-    # every other check still passes -- the medias load, the vectors are there,
-    # the patch grids are there, and the boxes are quietly pointing at the wrong
-    # pixels. Recomputing the band from the box is cheap and would have caught
-    # it at build time instead of via a human noticing a box drawn on snow.
-    for ds, emb in pc.cells():
-        if pc.DATASETS.get(ds, {}).get("kind") != "vg_scale":
-            continue
-        path = pc.cell_path(ds, emb)
-        if not path.exists():
-            continue
-        medias = io.load_medias(path)
-        bad = 0
-        checked = 0
-        for m in medias.values():
-            for cell in m.get("categories") or []:
-                boxes = [r["box"] for r in (m.get("regions") or []) if r.get("label") == cell]
-                if not boxes:
-                    continue
-                area = (max(b[2] for b in boxes) - min(b[0] for b in boxes)) * (
-                    max(b[3] for b in boxes) - min(b[1] for b in boxes)
-                )
-                lo, hi = pc.BOX_BANDS[cell.rsplit("@", 1)[1]]
-                checked += 1
-                if not (lo <= area < hi):
-                    bad += 1
-        if checked and bad:
-            problems.append(
-                f"{ds} x {emb}: {bad}/{checked} region boxes fall outside the band their cell "
-                f"name claims -- boxes and bands were measured in different pixel spaces"
-            )
-        # The band check above compares a box against its own label, so it is
-        # blind to a box corrupted BEFORE banding -- the band moves with it and
-        # the two stay consistent (#3281). This one compares the box against the
-        # frame, which nothing can drag along with it.
-        problems += [f"{ds} x {emb}: {g}" for g in region_geometry_problems(medias)]
-        break  # one embedder is enough; the boxes are identical across cells
-
-    # A derived cell that no longer matches its parent. `vg_scale_any` is a
-    # relabel of the built `vg_scale` pickle and shares its vectors, so it
-    # survives a parent rebuild looking perfect while carrying the parent's
-    # PREVIOUS labels, boxes and bands -- which is how a box repair ships to one
-    # study and not the other.
-    live_digest: dict[Path, str] = {}  # one parent serves every derived cell built from it
-    for ds, emb in pc.cells():
-        if pc.DATASETS.get(ds, {}).get("kind") != "vg_scale_any":
-            continue
-        path, parent = pc.cell_path(ds, emb), pc.cell_path("vg_scale", emb)
-        if not path.exists() or not parent.exists():
-            continue
-        medias = io.load_medias(path)
-        problems += [f"{ds} x {emb}: {g}" for g in region_geometry_problems(medias)]
-        first = next(iter(medias.values()), None)
-        stamped = ((first or {}).get("origin") or {}).get("params", {}).get("parent_labels")
-        if parent not in live_digest:
-            live_digest[parent] = scale_label_digest(io.load_medias(parent))
-        live = live_digest[parent]
-        if stamped is None:
-            problems.append(f"{ds} x {emb}: no parent_labels stamp -- built before the staleness check, rebuild it")
-        elif stamped != live:
-            problems.append(
-                f"{ds} x {emb}: derived from a {parent.name} that has since changed "
-                f"({stamped[:12]} != {live[:12]}) -- rebuild it, --force on the parent alone leaves it stale"
-            )
-
-    # A dataset's cells must all cover the same medias, or cross-embedder
-    # comparisons silently compare different populations. This is not
-    # hypothetical: a datadir missing its demo-source symlink sent the loader
-    # off to re-download the dataset, and it embedded a truncated 1662-media
-    # subset of a 4193-media dataset into a cell that otherwise looked healthy.
-    for ds, per_emb in counts_by_dataset.items():
-        if len(set(per_emb.values())) > 1:
-            majority = max(set(per_emb.values()), key=list(per_emb.values()).count)
-            odd = {e: n for e, n in per_emb.items() if n != majority}
-            problems.append(
-                f"{ds}: cells disagree on media count (most are {majority}); "
-                f"rebuild {', '.join(f'{e} ({n})' for e, n in sorted(odd.items()))}"
-            )
-
-    log(f"{'dataset':18s} {'embedder':14s} {'state':16s} {'medias':>7s} {'patch':>12s} {'dim':>6s}")
-    for ds, emb, state, n, patch, dim in rows:
-        log(f"{ds:18s} {emb:14s} {state:16s} {n:>7s} {patch:>12s} {dim:>6s}")
-
-    # Coverage is not implied by anything above: a cell can be structurally
-    # perfect and no longer contain the images a human reviewed. Reported here
-    # so a rebuild cannot be declared healthy without it being looked at.
-    try:
-        sys.path.insert(0, str(Path(__file__).resolve().parent))
-        import check_review_coverage  # noqa: PLC0415
-
-        if pc.cell_path("vg_scale", "siglip").exists():
-            log("")
-            log("review coverage:")
-            sys.argv = ["check_review_coverage"]
-            if check_review_coverage.main() != 0:
-                problems.append("vg_scale: the rebuild retired images that had been reviewed")
-    except Exception as exc:  # noqa: BLE001 - an absent review is not a build failure
-        log(f"  (review-coverage check skipped: {exc})")
-
-    if problems:
-        log("")
-        for p in problems:
-            log(f"PROBLEM: {p}")
-        return 1
-    log("all present cells verified")
-    return 0
-
-
-def _band_vocab_drift(dataset: str, chosen: list[str]) -> str:
-    """Would rebuilding this band reproduce the cells that already exist?
-
-    ``--rebuildable`` on its own answers a weaker question than it looks like it
-    answers: that selection *runs*, not that it selects the same thing. Those
-    come apart in the direction that hurts. #3297's two candidate repairs both
-    made the selector run again; only one of them kept picking the categories
-    the published ``vg_box_*`` sets were built from, and taking the other would
-    have silently redefined three datasets whose numbers are cited in #3129 and
-    #3156 -- with the right media count, the right vectors, and nothing to look
-    at that would say so.
-
-    So where a cell is already built, compare its vocabulary against what the
-    selector picks today. Reads the smallest present cell: every cell carries
-    ``categories``, so there is no reason to page in the multi-GB patch one.
-    Returns an empty string when they agree (or when nothing is built yet --
-    a purged pile has nothing to reproduce, which is not a failure).
-    """
-    present = [(pc.cell_path(dataset, e).stat().st_size, e) for e in pc.EMBEDDERS if pc.cell_path(dataset, e).exists()]
-    if not present:
-        return ""
-    _, emb = min(present)
-    medias = _cells_io().load_medias(pc.cell_path(dataset, emb))
-    live = {c for m in medias.values() for c in (m.get("categories") or [])}
-    gained = sorted(set(chosen) - live)
-    lost = sorted(live - set(chosen))
-    if not gained and not lost:
-        return ""
-    return (
-        f"{dataset}: a rebuild would NOT reproduce the built cells -- selection now differs "
-        f"from {dataset}__{emb}.pkl by {len(gained)} added and {len(lost)} dropped "
-        f"categories (added {gained[:5]}, dropped {lost[:5]}). That is a dataset change, "
-        f"not a rebuild; published numbers that cite this set would need re-examining."
-    )
-
-
-def rebuildable(datasets: list[str] | None = None) -> int:
-    """Exercise every dataset's *selection* step without embedding anything.
-
-    The pile documents itself as purgeable -- ``pile_config``: "every cell must
-    be rebuildable from sources that are **not** on scratch". Nothing checked
-    that. ``--verify`` loads the built cells, and a cell that loads says
-    nothing about whether it can be rebuilt: the two paths share no code, so a
-    rebuild path can rot for months behind a pile that verifies clean. It did
-    (#3297): a scan-format change on 2026-08-17 outran the scan file the
-    ``vg_box_*`` cells were selected from, and the break surfaced only when
-    somebody asked for a rebuild eleven days later.
-
-    So this is the canary that would have caught it the same day, for a few
-    seconds. It runs the part of each build that reads sources and decides
-    *what goes in the cell*, and skips the part that costs GPU-hours. For a
-    banded dataset that means really running :func:`_band_categories` and
-    reporting what it chose; elsewhere it means confirming the sources a
-    rebuild would read are present and in a shape the current code accepts.
-
-    Where a banded dataset is already built it goes one question further, via
-    :func:`_band_vocab_drift`: not only "would a rebuild run?" but "would a
-    rebuild produce *this*?". A repair that restores the former while quietly
-    changing the latter is the expensive kind, and it is invisible from the
-    built cell.
-
-    Deliberately does not parse the multi-GB sources (VG's ``objects.json``,
-    the COCO zip). A canary nobody runs is worth nothing, and the way to make
-    it run is to keep it cheap enough to sit in front of every build.
-    """
-    for bad in [d for d in (datasets or []) if d not in pc.DATASETS]:
-        raise SystemExit(f"unknown dataset {bad!r}; known: {sorted(pc.DATASETS)}")
-    wanted = list(datasets) if datasets else list(pc.DATASETS)
-
-    problems: list[str] = []
-
-    for ds in wanted:
-        spec = pc.DATASETS[ds]
-        kind = spec.get("kind")
-        try:
-            if kind == "demo":
-                pc.require_demo_source(ds)
-                log(f"  {ds:18s} ok       demo source staged")
-            elif kind == "coco":
-                # The *zip*, which is what `_load_coco` opens. Checking the
-                # extracted directory instead reported this dataset broken
-                # against sources that were entirely intact (#3299) -- a canary
-                # that names a different path than the build is not a canary.
-                for path in (pc.COCO_ANNOTATIONS, pc.COCO_VAL_ZIP):
-                    if not path.exists():
-                        raise SystemExit(f"{ds}: missing {path}")
-                log(f"  {ds:18s} ok       annotations + image zip present")
-            elif kind == "vg_band":
-                chosen = _band_categories(spec["band"])
-                drift = _band_vocab_drift(ds, chosen)
-                if drift:
-                    raise SystemExit(drift)
-                log(f"  {ds:18s} ok       {len(chosen)} categories selected")
-            elif kind == "vg_scale":
-                objects_json = pc.DEMO_CACHE / "visual_genome" / "objects.json"
-                if not objects_json.exists():
-                    raise SystemExit(f"{ds}: missing {objects_json}")
-                n_cells = len(pc.SCALE_CLASSES) * len(pc.BOX_BANDS)
-                roster = "roster present" if pc.ROSTER.exists() else "NO ROSTER (membership would be redrawn)"
-                log(f"  {ds:18s} ok       VG source present, {n_cells} cells, {roster}")
-            elif kind == "vg_scale_any":
-                # Derived from the built parent, so its "source" is a cell.
-                # Absent parents are not a problem here: a full run builds the
-                # parent first, and a purged pile rebuilds both in order.
-                built = [e for e in pc.EMBEDDERS if pc.cell_path("vg_scale", e).exists()]
-                log(f"  {ds:18s} ok       derives from vg_scale ({len(built)} parent cells built)")
-            else:
-                raise SystemExit(f"{ds}: unknown kind {kind!r}; this check needs teaching about it")
-        except SystemExit as exc:  # the loaders' own way of reporting a bad source
-            log(f"  {ds:18s} BROKEN   {exc}")
-            problems.append(f"{ds}: {exc}")
-
-    if problems:
-        log(f"{len(problems)} dataset(s) CANNOT be rebuilt from their sources:")
-        for p in problems:
-            log(f"REBUILD-BROKEN: {p}")
-        return 1
-    log("every dataset's selection step runs against its current sources")
-    return 0
-
-
-def report_bands() -> int:
-    """Report voted-box scale-band populations for each boxed dataset.
-
-    The bands are anchored to the patch embedder's geometry: ``sub_patch`` is
-    "smaller than one DINOv3 patch", i.e. below what the patch grid can resolve
-    at all. That anchoring is the band's whole meaning, so a thin ``sub_patch``
-    is a fact about the data, not a threshold to tune — widening the edge would
-    inflate the count with objects that *are* resolvable.
-
-    Reads the smallest available cell for each dataset: scale stats need only
-    ``regions``, which every cell carries, so there is no reason to page in the
-    multi-GB patch cell.
-    """
-    io = _cells_io()
-    cfg = _experiment_config()
-    from vtscore.eval.labels import category_scale_stats  # noqa: PLC0415
-
-    boxed = [ds for ds, info in pc.DATASETS.items() if info.get("boxed")]
-    if not boxed:
-        log("no boxed datasets in the pile; nothing to stratify")
-        return 0
-
-    for ds in boxed:
-        present = [(pc.cell_path(ds, e).stat().st_size, e) for e in pc.EMBEDDERS if pc.cell_path(ds, e).exists()]
-        if not present:
-            log(f"{ds}: no cells present")
-            continue
-        _, emb = min(present)
-        medias = io.load_medias(pc.cell_path(ds, emb))
-
-        counts: dict[str, int] = defaultdict(int)
-        for m in medias.values():
-            for c in m.get("categories") or [m.get("category")]:
-                if c:
-                    counts[c] += 1
-
-        selected, report = cfg.select_categories_by_scale(medias, dict(counts))
-        log("")
-        log(f"=== {ds}: {len(medias)} medias, {len(counts)} categories (via {emb}) ===")
-        dropped = report.get("dropped_above_max_voted_area") or []
-        log(f"  dropped above max_voted_area={report.get('max_voted_area')}: {len(dropped)}")
-        for name, info in (report.get("bands") or {}).items():
-            lo, hi = info["range"]
-            flag = "  ** UNDER-POPULATED **" if info["under_populated"] else ""
-            log(
-                f"  {name:14s} [{lo * 100:5.2f}%, {hi * 100:6.2f}%): "
-                f"{len(info['selected'])}/{info['target']} of {info['n_candidates']} candidates{flag}"
-            )
-            log(f"      {info['selected']}")
-
-        # When a band is starved, say whether the min-count filter is even the
-        # binding constraint. Measured on the first run it was not: the
-        # sub_patch pool held 5 categories (VG) and 1 (COCO) at every
-        # min_count from 5 to 30, so lowering it recovers nothing.
-        starved = [n for n, i in (report.get("bands") or {}).items() if i["under_populated"]]
-        if starved:
-            stats = {c: s for c in counts if (s := category_scale_stats(medias, c)) is not None}
-            for name in starved:
-                lo, hi = report["bands"][name]["range"]
-                pools = {
-                    mc: sum(1 for c, s in stats.items() if counts[c] >= mc and lo <= s["voted_area"] < hi)
-                    for mc in (5, 10, 20, 30)
-                }
-                spread = "same at every min_count" if len(set(pools.values())) == 1 else str(pools)
-                log(f"  {name}: candidate pool by min category count -> {spread}")
-        log(f"  -> selected {len(selected)} categories")
-    return 0
-
-
-def _sacct_build_nodes() -> dict[str, str]:
-    """dataset -> node, recovered from SLURM's accounting of the ``pile-*`` jobs.
-
-    Cells built before the sidecar existed are not anonymous after all: the build
-    ran as ``pile-<dataset>`` and ``sacct`` still knows which node took it. This
-    is recorded as ``hostname_recovered``, never as ``hostname`` -- it is an
-    inference from a job name, and one ambiguous dataset (two completed jobs) is
-    left out rather than guessed. It matters because the node determines the CPU,
-    and the CPU determines how the 384px resize rounds (#3160).
-    """
-    import subprocess  # noqa: PLC0415, S404 -- fixed argv, no shell
-
-    try:
-        out = subprocess.run(  # noqa: S603
-            ["sacct", "-X", "-S", "2026-01-01", "-n", "-P", "--format=JobName,NodeList,State"],  # noqa: S607
-            capture_output=True,
-            text=True,
-            timeout=60,
-            check=False,
-        )
-    except (OSError, subprocess.SubprocessError):
-        return {}
-    seen: dict[str, set[str]] = defaultdict(set)
-    for line in out.stdout.splitlines():
-        parts = line.split("|")
-        if len(parts) < 3 or not parts[0].startswith("pile-") or parts[2] != "COMPLETED":
-            continue
-        seen[parts[0][len("pile-") :]].add(parts[1])
-    return {ds: next(iter(nodes)) for ds, nodes in seen.items() if len(nodes) == 1}
-
-
-def provenance_report(backfill: bool = False) -> int:
-    """Show which device built each cell -- and, with ``--backfill-provenance``,
-    stamp what is still knowable for the cells built before this existed.
-
-    A backfilled sidecar deliberately records ``gpu_name: null``: the node a 2026
-    job ran on is not recoverable from the pickle, and writing a guess would be
-    worse than writing nothing. What it *can* record is the fingerprint, and that
-    is the half that matters for a rebuild -- it turns "did the rebuild reproduce
-    the cell?" from an unanswerable question into a hash comparison.
-    """
-    rows, missing, devices = [], [], defaultdict(list)
-    recovered = _sacct_build_nodes() if backfill else {}
-    for ds, emb in pc.cells():
-        cell = pc.cell_path(ds, emb)
-        if not cell.exists():
-            continue
-        path = pc.provenance_path(ds, emb)
-        if not path.exists():
-            if backfill:
-                stat = cell.stat()
-                record = {
-                    "dataset": ds,
-                    "embedder": emb,
-                    "cell": cell.name,
-                    "built_at": time.strftime("%Y-%m-%dT%H:%M:%S%z", time.localtime(stat.st_mtime)),
-                    "backfilled": True,
-                    "device": {
-                        "gpu_name": None,
-                        "hostname_recovered": recovered.get(ds),
-                        "recovered_from": "sacct pile-<dataset> job" if recovered.get(ds) else None,
-                        "note": "unknown: cell predates per-cell provenance (#3160)",
-                    },
-                    "cell_summary": {"megabytes": round(stat.st_size / 1e6, 1)},
-                    "fingerprint": cell_fingerprint(ds, emb),
-                }
-                path.write_text(json.dumps(record, indent=2) + "\n")
-                log(f"backfilled {path.name} ({record['fingerprint']['vectors_sha256'][:12]})")
-            else:
-                missing.append(f"{ds} x {emb}")
-                continue
-        rec = json.loads(path.read_text())
-        dev = rec.get("device", {})
-        if backfill and not dev.get("hostname") and not dev.get("hostname_recovered") and recovered.get(ds):
-            dev["hostname_recovered"] = recovered[ds]
-            dev["recovered_from"] = "sacct pile-<dataset> job"
-            rec["device"] = dev
-            path.write_text(json.dumps(rec, indent=2) + "\n")
-            log(f"recovered build node for {ds} x {emb}: {recovered[ds]}")
-        rows.append(
-            (
-                ds,
-                emb,
-                dev.get("gpu_name") or "unknown",
-                dev.get("hostname") or (f"{dev['hostname_recovered']}?" if dev.get("hostname_recovered") else "-"),
-                str(dev.get("cpu_capability") or "-"),
-                (dev.get("commit") or "-")[:9],
-                rec.get("fingerprint", {}).get("vectors_sha256", "")[:12],
-            )
-        )
-        devices[(dev.get("gpu_name") or "unknown", dev.get("cpu_capability"))].append(f"{ds}x{emb}")
-
-    log(f"{'dataset':<18} {'embedder':<14} {'device':<26} {'node':<10} {'dispatch':<9} {'commit':<10} vectors")
-    for row in sorted(rows):
-        log("{:<18} {:<14} {:<26} {:<10} {:<9} {:<10} {}".format(*row))
-    if missing:
-        log(f"\n{len(missing)} cell(s) with NO provenance (run --backfill-provenance): {', '.join(missing)}")
-    if len(devices) > 1:
-        log(f"\nthis pile MIXES {len(devices)} build environments. The measured cost of mixing")
-        log("hosts is 1.5e-04 median 1-cos on siglip2_l when CPU dispatch is unpinned (#3160):")
-        for (name, cap), cells in sorted(devices.items(), key=lambda kv: str(kv[0])):
-            log(f"  {str(name):<26} dispatch={cap or 'unrecorded':<10} {len(cells)} cell(s)")
-    return 0
-
-
-def _manifest_provenance(dataset: str, embedder: str) -> dict:
-    """The provenance fields the manifest carries per cell, or nulls if unknown."""
-    path = pc.provenance_path(dataset, embedder)
-    if not path.exists():
-        return {"gpu_name": None, "built_by": None, "commit": None, "vectors_sha256": None}
-    rec = json.loads(path.read_text())
-    dev = rec.get("device", {})
-    return {
-        "gpu_name": dev.get("gpu_name"),
-        "built_by": dev.get("hostname"),
-        "commit": dev.get("commit"),
-        "vectors_sha256": rec.get("fingerprint", {}).get("vectors_sha256"),
-    }
-
-
-def write_manifest() -> None:
-    """Write MANIFEST.json + MANIFEST.md describing the pile and how to rebuild it."""
-    io = _cells_io()
-    entries = []
-    for ds, emb in pc.cells():
-        path = pc.cell_path(ds, emb)
-        if not path.exists():
-            entries.append({"dataset": ds, "embedder": emb, "present": False})
-            continue
-        medias = io.load_medias(path)
-        n = len(medias)
-        entries.append(
-            {
-                "dataset": ds,
-                "embedder": emb,
-                "present": True,
-                "file": path.name,
-                "megabytes": round(path.stat().st_size / 1e6, 1),
-                "n_medias": n,
-                "n_patch_grids": sum(1 for m in medias.values() if m.get("patch_grid") is not None),
-                "region_capable": pc.region_capable(ds, emb),
-                # Which machine built it (#3160). None for cells that predate the
-                # sidecar; a null here is a fact about the pile, not a gap to hide.
-                **_manifest_provenance(ds, emb),
-            }
-        )
-
-    doc = {
-        "pile": str(pc.PILE),
-        "sources": {
-            "demo_cache": str(pc.DEMO_CACHE),
-            "coco_root": str(pc.COCO_ROOT),
-        },
-        "datasets": pc.DATASETS,
-        "embedders": pc.EMBEDDERS,
-        "cells": entries,
-    }
-    (pc.PILE / "MANIFEST.json").write_text(json.dumps(doc, indent=2) + "\n")
-
-    present = [e for e in entries if e["present"]]
-    total_mb = sum(e["megabytes"] for e in present)
-    lines = [
-        "# Pre-embedded pile",
-        "",
-        f"`{pc.PILE}` — {len(present)}/{len(entries)} cells, {total_mb / 1000:.1f} GB of embeddings.",
-        "",
-        "Point a study at it with:",
-        "",
-        "```bash",
-        f'export VTSEARCH_DATA_DIR="{pc.DATADIR}"',
-        f'export VTSEARCH_MODELS_DIR="{pc.MODELS}"',
-        "```",
-        "",
-        "## Cells",
-        "",
-        "| dataset | embedder | medias | patch grids | region-voting | size |",
-        "|---|---|---:|---:|:--:|---:|",
-    ]
-    for e in entries:
-        if not e["present"]:
-            lines.append(f"| `{e['dataset']}` | `{e['embedder']}` | — | — | — | *missing* |")
-            continue
-        region = "**yes**" if e["region_capable"] else "no"
-        lines.append(
-            f"| `{e['dataset']}` | `{e['embedder']}` | {e['n_medias']} | "
-            f"{e['n_patch_grids']} | {region} | {e['megabytes']:.0f} MB |"
-        )
-    lines += [
-        "",
-        "**Region voting needs both halves**: ground-truth boxes (dataset) *and* a patch",
-        "grid (embedder). A boxed dataset on a single-vector embedder silently runs as",
-        "binary voting — the failure behind #2877, #2897 and #2905. `build_pile.py --verify`",
-        "asserts the geometry rather than trusting the arm table.",
-        "",
-        "## Rebuilding",
-        "",
-        "Scratch is treated as purgeable. Every cell rebuilds from staged, non-scratch",
-        "sources, so the pile is disposable:",
-        "",
-        "```bash",
-        "python build_pile.py            # rebuild whatever is missing (idempotent)",
-        "python build_pile.py --verify   # check geometry after a rebuild",
-        "```",
-        "",
-        f"Sources: demo datasets from `{pc.DEMO_CACHE}`, COCO from `{pc.COCO_ROOT}`.",
-        "",
-    ]
-    (pc.PILE / "MANIFEST.md").write_text("\n".join(lines))
-    log(f"wrote MANIFEST.json + MANIFEST.md ({len(present)}/{len(entries)} cells)")
-
-
-def list_cells() -> None:
-    log(f"pile: {pc.PILE}")
-    for ds, emb in pc.cells():
-        path = pc.cell_path(ds, emb)
-        mark = "present" if path.exists() else "MISSING"
-        size = f"{path.stat().st_size / 1e6:8.0f} MB" if path.exists() else " " * 11
-        region = " region-voting" if pc.region_capable(ds, emb) else ""
-        log(f"  {ds:18s} x {emb:14s} {mark:8s} {size}{region}")
 
 
 def main() -> int:
